backend/requirements.txt
Flask
Flask-CORS
python-dotenv
groq
googlemaps